requirements.txt
faiss-cpu
selenium
beautifulsoup
sentence-transformers
openai
numpy
json